Move intdup to qthelper
